ground up rearchitecting of nuxt.config for nuxt/content move
