--- content/blog/diana-rodriguez-joins-the-developer-relations-team.md
@@ -1,26 +1,0 @@
----
-title: Diana Rodriguez Joins the Developer Relations Team
-description: Diana Rodriguez is joining the Developer Relations Team as Python
-  Developer Advocate. We are very excited to have her on board!
-thumbnail: /content/blog/diana-rodriguez-joins-the-developer-relations-team/thumbnail.jpg
-author: dianarodriguez
-published: true
-published_at: 2020-03-19T18:52:39.194Z
-comments: false
-category: team
-tags:
-  - python
----
-## Hola Mundo!!/Hello World!
-
-I am definitely excited to join the Developer Relations team!
-
-I live in the “Bull City”, aka Durham, North Carolina but I am a citizen of the world. My background is super diverse. I grew up in Venezuela and lived in many countries throughout my life.
-
-I am a polyglot on all counts, not only programming languages! I speak Spanish, English, Portuguese, Italian, French, Thai, Mandarin, Swedish, Norwegian, Khmer, Lao, Malay and sign language. All the space in my head is containerized with Kubernetes and I have three pods, one of them for languages, another one for programming and the last one for music.
-
-My background is mainly backend development with Python and infrastructure. I am passionate about DevOps and fostering collaboration between all sectors of the tech ecosystem and, especially advocating for diversity and inclusion. I believe that there is a place for everyone to sit at the table.
-
-When I am not nerding out, I like to play video games and/or board games. I am a big fan of rugby and your favourite #11 if I get to play. I have a black belt in Taekwondo and I am also a big fan of dining out and trying exciting food!
-
-I look forward to engaging with our amazing community and all the goodness that are coming from that. Awesome things are in store for 2020 and the future!
