--- content/blog/greg-holmes-joins-the-vonage-developer-education-team.md
@@ -1,28 +1,0 @@
----
-title: Greg Holmes Joins the Vonage Developer Education Team
-description: Greg Holmes joins the Vonage Developer Education team as a
-  Developer Educator. We couldn't be more thrilled!
-thumbnail: /content/blog/greg-holmes-joins-the-vonage-developer-education-team/greg_1200_x_600_1.png
-author: gregholmes
-published: true
-published_at: 2020-03-24T19:06:41.798Z
-comments: false
-category: team
-tags:
-  - php
----
-Hi! My name is Greg Holmes, and I’m excited to be joining the Vonage team as a Developer Educator. For the last two months, I have been painfully containing my excitement while working my notice period for my former employer. But, now the time has come!
-
-I am originally from London but now live in a little village near Colchester called Thorpe-le-Soken, which is surrounded by fields, farmland, and very close to the coast. This part of Essex was where the river Thames [opened to the sea around 500,000 years ago](http://www.geoessex.org.uk/the_early_ice_age.html). This makes it a perfect place to find fossilised shark teeth released from the red crag by the sea.
-
-I have been involved in tech one way or another for most of my life, having studied various courses at college including a City & Guilds in Visual Basic 6, from college moved on to Anglia Ruskin University where I studied Computer and Information Systems. My family has always seen me as the go-to person for their computer problems (often to fix a printer).
-
-Before joining Vonage, I spent 9 years working at two other companies. I started my career as a PHP developer at HMA for around 4 1/2 years before moving to a similar role for the online estate agent Housesimple. After these 9 years of working within closed source companies, taking from the developer community, my hope is that I can now give back to the community.
-
-As you can guess, my primary language of choice is PHP. However, I am more than happy to dive into any language that suits the needs of my project.
-
-In my spare time, I’m always out and about, or busy at home. I enjoy climbing, gaming, and creating whatever development project takes my fancy. Quite often this will involve a Raspberry Pi or two.
-
-I’m very excited to be part of this team and company. I can’t wait to be able to give back to the community and have a lot of fun doing so!
-
-You can contact me on the Vonage Community Slack, or on Twitter [@greg__holmes](https://twitter.com/Greg__Holmes).
