+++ package.json
@@ -13,20 +13,15 @@
     "lintfix": "eslint --fix --ext .js,.vue --ignore-path .gitignore ."
   },
   "dependencies": {
+    "@nuxt/content": "^1.2.0",
     "@nuxt/http": "^0.3.0",
     "@nuxtjs/dotenv": "^1.4.1",
     "@nuxtjs/feed": "^1.1.0",
-    "@toycode/markdown-it-class": "^1.2.3",
     "@vonagevolta/volta2": "0.0.37",
     "algoliasearch": "^4.2.0",
-    "front-matter": "^3.1.0",
-    "frontmatter-markdown-loader": "3.1.0",
-    "html-to-text": "^5.1.1",
+    "defu": "^2.0.4",
+    "front-matter": "^4.0.2",
     "instantsearch.css": "^7.4.2",
-    "markdown-it": "^10.0.0",
-    "markdown-it-anchor": "^5.2.7",
-    "markdown-it-front-matter": "^0.2.1",
-    "markdown-it-prism": "^2.0.5",
     "nuxt": "^2.5.1",
     "vue-disqus": "^3.0.5",
     "vue-fragment": "^1.5.1",
