+++ pages/authors/index.vue
@@ -1,6 +1,6 @@
 <template>
   <section class="Blog__Full-width">
-    <header class="Blog__Full-width">
+    <!-- <header class="Blog__Full-width">
       <AuthorHero />
     </header>
     <main class="Vlt-container">
@@ -18,45 +18,45 @@
           type="card"
         />
       </div>
-    </main>
+    </main> -->
   </section>
 </template>
 
 <script>
-import AuthorHero from "~/components/AuthorHero"
-import Author from "~/components/Author"
-import Breadcrumbs from "~/components/Breadcrumbs"
+// import AuthorHero from "~/components/AuthorHero"
+// import Author from "~/components/Author"
+// import Breadcrumbs from "~/components/Breadcrumbs"
 
-export default {
-  components: {
-    Breadcrumbs,
-    AuthorHero,
-    Author,
-  },
+// export default {
+//   components: {
+//     Breadcrumbs,
+//     AuthorHero,
+//     Author,
+//   },
 
-  data() {
-    const { authors } = require("~/content/authors.json")
+//   data() {
+//     const { authors } = require("~/content/authors.json")
 
-    authors.sort((a, b) => {
-      const nameA = a.name.toUpperCase()
-      const nameB = b.name.toUpperCase()
-      return (nameA < nameB) ? -1 : (nameA > nameB) ? 1 : 0
-    })
+//     authors.sort((a, b) => {
+//       const nameA = a.name.toUpperCase()
+//       const nameB = b.name.toUpperCase()
+//       return (nameA < nameB) ? -1 : (nameA > nameB) ? 1 : 0
+//     })
 
-    return {
-      authors: authors,
-      routes: [
-        { route: `/authors`, title: `All our authors`, current: true },
-      ]
-    }
-  },
+//     return {
+//       authors: authors,
+//       routes: [
+//         { route: `/authors`, title: `All our authors`, current: true },
+//       ]
+//     }
+//   },
 
-  head() {
-    return {
-      title: `All the amazing people who contribute to our content`
-    }
-  },
-}
+//   head() {
+//     return {
+//       title: `All the amazing people who contribute to our content`
+//     }
+//   },
+// }
 </script>
 
 <style scoped>
