+++ pages/index.vue
@@ -1,6 +1,6 @@
 <template>
   <section class="Blog__Full-width">
-    <header class="Blog__Full-width">
+    <!-- <header class="Blog__Full-width">
       <SearchHero />
     </header>
     <main class="Vlt-container">
@@ -35,81 +35,81 @@
       <NLink to="/archive" no-prefetch class="Vlt-btn Vlt-btn--quaternary">
         View older posts
       </NLink>
-    </footer>
+    </footer> -->
   </section>
 </template>
 
 <script>
-import SearchHero from "~/components/SearchHero"
-import Card from "~/components/Card"
-import MiniCard from "~/components/MiniCard"
+// import SearchHero from "~/components/SearchHero"
+// import Card from "~/components/Card"
+// import MiniCard from "~/components/MiniCard"
 
-export default {
-  components: {
-    Card,
-    MiniCard,
-    SearchHero,
-  },
+// export default {
+//   components: {
+//     Card,
+//     MiniCard,
+//     SearchHero,
+//   },
 
-  data() {
-    return {
-      featuredPosts: this.getFeaturedPosts(), // limit to 2 "latest posts"
-      categories: this.getCategories(),
-      posts: this.getPublishedPosts().slice(0, 6), // limit to 6 "latest posts"
-    }
-  },
+//   data() {
+//     return {
+//       featuredPosts: this.getFeaturedPosts(), // limit to 2 "latest posts"
+//       categories: this.getCategories(),
+//       posts: this.getPublishedPosts().slice(0, 6), // limit to 6 "latest posts"
+//     }
+//   },
 
-  methods: {
-    getPublishedPosts() {
-      const resolve = require.context("~/content/", true, /\.md$/)
-      const imports = resolve
-        .keys()
-        .map((key) => {
-          const [, name] = key.match(/\/(.+)\.md$/) // eslint-disable-line no-unused-vars
-          return resolve(key)
-        })
-        .filter((content) => content.attributes.published != false)
+//   methods: {
+//     getPublishedPosts() {
+//       const resolve = require.context("~/content/", true, /\.md$/)
+//       const imports = resolve
+//         .keys()
+//         .map((key) => {
+//           const [, name] = key.match(/\/(.+)\.md$/) // eslint-disable-line no-unused-vars
+//           return resolve(key)
+//         })
+//         .filter((content) => content.attributes.published != false)
 
-      imports.sort((a, b) => {
-        const aPublishedDate = new Date(a.attributes.published_at)
-        const bPublishedDate = new Date(b.attributes.published_at)
-        return bPublishedDate - aPublishedDate
-      })
+//       imports.sort((a, b) => {
+//         const aPublishedDate = new Date(a.attributes.published_at)
+//         const bPublishedDate = new Date(b.attributes.published_at)
+//         return bPublishedDate - aPublishedDate
+//       })
 
-      return imports
-    },
+//       return imports
+//     },
 
-    getCategories() {
-      const posts = this.getPublishedPosts()
+//     getCategories() {
+//       const posts = this.getPublishedPosts()
 
-      return Array.from(new Set(posts.map((post) => post.attributes.category)))
-    },
+//       return Array.from(new Set(posts.map((post) => post.attributes.category)))
+//     },
 
-    getFeaturedPosts() {
-      return this.getPublishedPosts().slice(0, 2)
-    },
+//     getFeaturedPosts() {
+//       return this.getPublishedPosts().slice(0, 2)
+//     },
 
-    categorySizeMap(category) {
-      if (category ==='tutorial') { 
-        return 6
-      } else {
-        return 3
-      }
-    },
+//     categorySizeMap(category) {
+//       if (category ==='tutorial') { 
+//         return 6
+//       } else {
+//         return 3
+//       }
+//     },
 
-    getCategoryPosts(category) {
-      return this.getPublishedPosts()
-        .filter((post) => post.attributes.category === category)
-        .slice(0, this.categorySizeMap(category))
-    }
-  },
+//     getCategoryPosts(category) {
+//       return this.getPublishedPosts()
+//         .filter((post) => post.attributes.category === category)
+//         .slice(0, this.categorySizeMap(category))
+//     }
+//   },
 
-  head() {
-    return {
-      title: "We ♥ content"
-    }
-  },
-}
+//   head() {
+//     return {
+//       title: "We ♥ content"
+//     }
+//   },
+// }
 </script>
 
 <style scoped>
